♻️ (Commands) - Simplify commands by creating 2 commands that group functions by category : setup and docs
