+++ l10n/bundle.l10n.fr.json
@@ -2,15 +2,30 @@
   "Install": "Installer",
   "Enable": "Activer",
   "Download": "Télécharger",
+
+  "🛠️ Select a setup option": "🛠️ Sélectionner une option de configuration",
+  "🧩 Install Python extension": "🧩 Installer l'extension Python",
+  "📕 Install Revit API Stubs": "📕 Installer les Stubs Revit API",
+  "📕 Install pyRevit": "📕 Installer pyRevit",
+  "📝 Update \"settings.json\"": "📝 Mettre à jour \"settings.json\"",
+
+  "📚 Select a documentation": "📚 Sélectionner une documentation",
+  "📕 Revit API Docs": "📕 Revit API Docs",
+  "📕 pyRevit": "📕 pyRevit",
+
+  "pyRevit-with-vscode only supports Windows operating system.": "pyRevit-with-vscode n'est compatible qu'avec le système d'exploitation Windows",
+
   "The Python extension is required to use pyRevit-with-vscode. Do you want to install or enable it now?": "L'extension Python est requise pour utiliser pyRevit-with-vscode. Voulez-vous l'installer ou l'activer maintenant ?",
   "The Python extension is already installed and activated.": "L'extension Python est déjà installée et activée.",
-  "pyRevit-with-vscode only supports Windows operating system.": "pyRevit-with-vscode n'est compatible qu'avec le système d'exploitation Windows",
+
   "The Revit API Stubs are required to use pyRevit-with-vscode. Do you want to download it now?": "Les Revit API Stubs sont requises pour utiliser pyRevit-with-vscode. Voulez-vous les télécharger maintenant ?",
   "The Revit API Stubs are download and place in %APPDATA%.": "Les Revit API Stubs sont téléchargés et placés dans %APPDATA%.",
   "The Revit API Stubs are already installed in %APPDATA%.": "Les Revit API Stubs sont déjà installés dans %APPDATA%.",
+
   "pyRevit is required to use pyRevit-with-vscode. Do you want to download it now?": "pyRevit est nécessaire pour utiliser pyRevit-with-vscode. Voulez-vous le télécharger maintenant ?",
   "pyRevit is already installed in %APPDATA%.": "pyRevit est déjà installé dans %APPDATA%.",
-  "\"settings.json\" has been updated.": "\"settings.json\" a été mis à jour.",
+
   "Select the Revit API versions": "Sélectionnez les versions de Revit API",
+  "\"settings.json\" has been updated.": "\"settings.json\" a été mis à jour.",
   "Revit API versions selection aborted": "La sélection des versions de Revit API a été annulée"
 }
